Stage

Former-commit-id: ef3c6de13626b308287d4d1455824926aef24302

+++ Source/Python/datalabs/etl/orm/load.py
@@ -17,6 +17,7 @@ LOGGER = logging.getLogger(__name__)
 LOGGER.setLevel(logging.DEBUG)
 
 
+# pylint: disable=too-many-instance-attributes
 @dataclass
 class TableParameters:
     data: str
@@ -24,6 +25,7 @@ class TableParameters:
     model_class: str
     primary_key: str
     columns: list
+    schema: str
     current_hashes: pandas.DataFrame
     incoming_hashes: pandas.DataFrame
 
@@ -35,7 +37,7 @@ class ORMLoaderTask(LoaderTask, DatabaseTaskMixin):
         with self._get_database(Database, self._parameters) as database:
             for model_class, data, table in zip(self._get_model_classes(),
                                                 self._get_dataframes(),
-                                                self._parameters['TABLES']):
+                                                self._parameters['TABLES'].split(',')):
 
                 table_parameters = self._generate_table_parameters(model_class, data, table, database)
                 self._update(database, table_parameters)
@@ -44,13 +46,16 @@ class ORMLoaderTask(LoaderTask, DatabaseTaskMixin):
             database.commit()
 
     def _generate_table_parameters(self, model_class, data, table, database):
-        primary_key = self._get_primary_key(database, table)
-        columns = self._get_database_columns(database, table)
+        schema = self._parameters['SCHEMA']
+        primary_key = self._get_primary_key(database, table, schema)
+        columns = self._get_database_columns(database, table, schema)
 
-        current_hashes = self._get_current_row_hashes(database, table, primary_key)
+        data[primary_key] = [str(key) for key in data[primary_key]]
+
+        current_hashes = self._get_current_row_hashes(database, table, schema, primary_key)
         incoming_hashes = self._generate_row_hashes(columns, data, primary_key)
 
-        return TableParameters(data, table, model_class, primary_key, columns, current_hashes, incoming_hashes)
+        return TableParameters(data, table, model_class, primary_key, columns, schema, current_hashes, incoming_hashes)
 
     def _update(self, database, table_parameters):
         self._add_data(database, table_parameters)
@@ -58,38 +63,51 @@ class ORMLoaderTask(LoaderTask, DatabaseTaskMixin):
         self._update_data(database, table_parameters)
 
     @classmethod
-    def _get_primary_key(cls, database, table):
+    def _get_primary_key(cls, database, table, schema):
         query = "SELECT a.attname, format_type(a.atttypid, a.atttypmod) AS data_type FROM pg_index i " \
                 "JOIN pg_attribute a ON a.attrelid = i.indrelid AND a.attnum = ANY(i.indkey) " \
-                f"WHERE  i.indrelid = f'{table}'::regclass AND i.indisprimary"
+                f"WHERE  i.indrelid = '{schema}.{table}'::regclass AND i.indisprimary"
 
-        primary_key_table = pandas.read_sql(query, database)
+        primary_key_table = database.read(query)
 
         return primary_key_table['attname'][0]
 
     @classmethod
-    def _get_database_columns(cls, database, table):
+    def _get_database_columns(cls, database, table, schema):
         query = "SELECT * FROM information_schema.columns " \
-                f"WHERE table_schema = 'oneview' AND table_name = f'{table}';"
-        old_data = pandas.read_sql(query, database)
+                f"WHERE table_schema = '{schema}' AND table_name = '{table}';"
+        old_data = database.read(query)
 
-        return old_data.columns
+        return old_data.column_name.to_list()
 
     @classmethod
-    def _get_current_row_hashes(cls, database, table, primary_key):
-        get_current_hash = f"SELECT f'{primary_key}', md5(f'{table}'::TEXT) FROM f'{table}'"
+    def _get_current_row_hashes(cls, database, table, schema, primary_key):
+        get_current_hash = f"SELECT {primary_key}, md5({table}::TEXT) FROM {schema}.{table}"
 
-        return pandas.read_sql(get_current_hash, database)
+        return database.read(get_current_hash)
 
     @classmethod
     def _generate_row_hashes(cls, columns, data, primary_key):
         csv_data = data[columns].to_csv(header=None, index=False).strip('\n').split('\n')
-        row_strings = ["(" + i + ")" for i in csv_data]
+        row_strings = ["(" + cls.add_quotes(i) + ")" for i in csv_data]
 
         hashes = [hashlib.md5(row_string.encode('utf-8')).hexdigest() for row_string in row_strings]
         primary_keys = data[primary_key].tolist()
 
-        return pandas.DataFrame({primary_key: primary_keys, 'md5': hashes})
+        incoming_hashes = pandas.DataFrame({primary_key: primary_keys, 'md5': hashes})
+
+        return incoming_hashes
+
+    @classmethod
+    def add_quotes(cls, csv_string):
+        conditions = [',', ' ']
+        string_list = csv_string.split(',')
+
+        if any(string_list[1].find(c) >= 0 for c in conditions):
+            string_list[1] = '"' + string_list[1] + '"'
+            csv_string = ','.join(string_list)
+
+        return csv_string
 
     @classmethod
     def _add_data(cls, database, table_parameters):
@@ -107,11 +125,12 @@ class ORMLoaderTask(LoaderTask, DatabaseTaskMixin):
 
     @classmethod
     def _add_data_to_table(cls, table_parameters, data, database):
-        models = [cls._create_model(table_parameters.model_class, row) for row in data.itertuples(index=False)]
+        if not data.empty:
+            models = [cls._create_model(table_parameters.model_class, row) for row in data.itertuples(index=False)]
 
-        for model in models:
-            # pylint: disable=no-member
-            database.add(model)
+            for model in models:
+                # pylint: disable=no-member
+                database.add(model)
 
     @classmethod
     def _delete_data(cls, database, table_parameters):
@@ -129,16 +148,25 @@ class ORMLoaderTask(LoaderTask, DatabaseTaskMixin):
 
     @classmethod
     def _delete_data_from_table(cls, table_parameters, data, database):
-        deleted_primary_keys = data[table_parameters.primary_key].tolist()
-        database_rows_query = f"SELECT * FROM {table_parameters.table} " \
-                              f"WHERE {table_parameters.primary_key} IN {tuple(deleted_primary_keys)};"
+        if not data.empty:
+            deleted_primary_keys = data[table_parameters.primary_key].tolist()
+            database_rows_query = "SELECT * FROM {}.{} WHERE {} IN ({});".format(table_parameters.schema,
+                                                                                 table_parameters.table,
+                                                                                 table_parameters.primary_key,
+                                                                                 ",".join(["'{}'".format(x)
+                                                                                           for x in deleted_primary_keys
+                                                                                           ]
+                                                                                          )
+                                                                                 )
 
-        deleted_data = pandas.read_sql(database_rows_query, database)
-        models = [cls._create_model(table_parameters.model_class, row) for row in deleted_data.itertuples(index=False)]
+            deleted_data = database.read(database_rows_query)
+            models = [cls._create_model(table_parameters.model_class, row) for row in
+                      deleted_data.itertuples(index=False)
+                      ]
 
-        for model in models:
-            # pylint: disable=no-member
-            database.delete(model)
+            for model in models:
+                # pylint: disable=no-member
+                database.delete(model)
 
     @classmethod
     def _update_data(cls, database, table_parameters):
@@ -171,10 +199,11 @@ class ORMLoaderTask(LoaderTask, DatabaseTaskMixin):
 
     @classmethod
     def _update_data_in_table(cls, table_parameters, data, database):
-        models = [cls._create_model(table_parameters.model_class, row) for row in data.itertuples(index=False)]
+        if not data.empty:
+            models = [cls._create_model(table_parameters.model_class, row) for row in data.itertuples(index=False)]
 
-        for model in models:
-            database.query().update(model)
+            for model in models:
+                database.update(model)
 
     def _get_model_classes(self):
         return [import_plugin(table) for table in self._parameters['MODEL_CLASSES'].split(',')]
